pyproject.toml
[project]
name = "dbx-ds-monorepo"
version = "0.0.0"
requires-python = ">=3.12"
dependencies = []

[dependency-groups]
# uv sync installs the `dev` group by default, which makes onboarding much easier.
dev = [
  "ruff>=0.5",
  "pytest>=8",
  "mypy>=1.10",
  "pre-commit>=3.7",
  "typer>=0.12",
  "pyyaml>=6.0",
]

[tool.uv.workspace]
members = ["packages/*", "projects/*"]

[tool.ruff]
line-length = 100
target-version = "py312"
src = ["packages", "projects"]

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "SIM"]
ignore = ["E501"]

[tool.pytest.ini_options]
addopts = "-q"
testpaths = ["packages", "projects"]
markers = [
  "integration: tests intended to run on Databricks compute",
]

[tool.mypy]
python_version = "3.12"
warn_unused_configs = true
warn_return_any = true
ignore_missing_imports = true
